Improved scraper The scraper now checks for duplicate entries by checking a dictionary in memory instead of checking the database for every new entry.
Scraper also loads the most recent entry for every bus from the database
into memory to prevent duplicate entries upon restarting it. I also
moved the place of the API key variable just to keep it close to the
only other changing variable, the location of the database. Finally, the
database file is now at the root of the repo.